# foss_public_alert_server/subscriptionHandler/tests.py
# SPDX-FileCopyrightText: Nucleus <nucleus-ffm@posteo.de>
# SPDX-License-Identifier: AGPL-3.0-or-later

import datetime
import json
import logging
import requests

from django.contrib.gis.geos import Polygon
from django.test import TestCase
from django.test import Client

from alertHandler.models import Alert
from .models import Subscription, ConnectionFlag
from .tasks import remove_old_subscription

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class SubscriptionHandlerTestsCase(TestCase):
    fixtures = ["appSettingsDump.json"]

    client = Client()

    def test_successful_subscription(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertContains(response, 'successfully subscribed', status_code=200)

    def test_subscription_without_user_agent(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json")
        self.assertContains(response, 'successfully subscribed', status_code=200)

    def test_successful_unsubscribe(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json")
        data = json.loads(response.content)
        subscription_id = data["subscription_id"]

        response = self.client.delete(f'/subscription/?subscription_id={subscription_id}', content_type="application/json")
        self.assertContains(response, 'successfully unsubscribed', status_code=200)

    def test_unsuccessful_unsubscribe_invalid_subscription_id(self):
        response = self.client.delete('/subscription/?subscription_id=invalid_id', content_type="application/json")
        self.assertContains(response, 'invalid subscription id', status_code=400)

    def test_push_service_not_supported(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "EXAMPLE_PUSH",
            'token': 'https://example.com'
        }
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json")
        self.assertContains(response, b'push service is not available on this instance.', status_code=400)

    def test_blocked_unifiedPush(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://ntfy.sh/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', data, content_type="application/json")
        self.assertContains(response, b'Your UnifiedPush Server ntfy.sh is blocked. We can not reliably deliver push notifications to this server. Please choose another one.', status_code=400)

    def test_antimeridian_cross(self):
        data = {
            'min_lat': 179.0,
            'max_lat': -179.0,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', data, content_type="application/json")
        self.assertContains(response, b'invalid bounding box', status_code=400)

    def test_invalid_parameters(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            #'min_lon': 8.591, missing
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', data, content_type="application/json")
        self.assertContains(response, b'invalid or missing parameters', status_code=400)

    def test_non_functional_push_service(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            'p256dh_key': 'BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0',
            'auth_key': 'ns9swjbbKTEN12VGW_tJqA',
        }
        urls = [
            'https://alerts.kde.org/non-existing-content',  # 404 (subscription expired)
            'https://non.existing.domain.tld/',  # name resolution failure
            'https://85.215.55.234/non-existing-content',  # TLS certification mismatch
            'https://429.returnco.de/whatever',  # rate limited
            'https://500.returnco.de/whatever',  # 500
            # TODO timeouting endpoint
        ]
        flag_count = ConnectionFlag.objects.count()
        for url in urls:
            for service in ['UNIFIED_PUSH', 'UNIFIED_PUSH_ENCRYPTED']:
                data['push_service'] = service
                data['token'] = url
                response = self.client.post('/subscription/', data, content_type="application/json")
                self.assertEqual(response.status_code, 400)
        self.assertEqual(flag_count, ConnectionFlag.objects.count())

    def test_send_notification(self):
        for alert in Alert.objects.all():
            for subscription in Subscription.objects.filter(bounding_box__intersects=alert.area):
                logger.debug(f"Send notification for {subscription.id} to {subscription.distributor_url}")
                requests.post(subscription.token, json.dumps(alert.alert_id)) #@todo fix
        # @todo check performance

    def test_update_subscription_sent_heartbeat(self):
        # create subscription
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            "push_service": "UNIFIED_PUSH",
            'token': 'https://unifiedpush.kde.org/J9gTXxwbOEKNfeJW'
        }
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        data = json.loads(response.content)
        subscription_id = data["subscription_id"]
        # update subscription
        response = self.client.put(f'/subscription/?subscription_id={subscription_id}', headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertContains(response, 'Subscription successfully updated', status_code=200)

    def test_update_subscription_invalid_subscription_id(self):
        response = self.client.put(f'/subscription/?subscription_id=invalid',
                                   content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertEqual(response.status_code, 404)

    def test_update_subscription_old_subscription_id(self):
        response = self.client.put(f'/subscription/?subscription_id=e1ce46fb-a885-4b26-5ba8-708cccfcfa2b',
                                   content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertContains(response, 'Subscription has expired. You must register again!', status_code=404)

    def test_invalid_up_token(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            'p256dh_key': 'BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0',
            'auth_key': 'ns9swjbbKTEN12VGW_tJqA',
        }
        invalid_tokens = [
            "",
            "127.0.0.1",
            "http://unfied.push.org",
            "https://192.168.178.42",
            "https://localhost:1234",
            "https://ntfy.sh/J9gTXxwbOEKNfeJW&up=1"
        ]

        # invalid tokens on initial subscription
        for token in invalid_tokens:
            data["token"] = token
            data["push_service"] = "UNIFIED_PUSH"
            response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
            self.assertEqual(response.status_code, 400)
        for token in invalid_tokens:
            data["token"] = token
            data["push_service"] = "UNIFIED_PUSH_ENCRYPTED"
            response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
            self.assertEqual(response.status_code, 400)

        # invalid tokens on subscription updates
        data["token"] = "https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1"
        data["push_service"] = "UNIFIED_PUSH"
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertEqual(response.status_code, 200)
        sub_id = response.json()["subscription_id"]
        self.assertIsNotNone(sub_id)

        data["subscription_id"] = sub_id
        for token in invalid_tokens:
            data["token"] = token
            data["push_service"] = "UNIFIED_PUSH"
            response = self.client.put(f'/subscription/?subscription_id={sub_id}', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
            self.assertEqual(response.status_code, 400)

        data["token"] = "https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1"
        data["push_service"] = "UNIFIED_PUSH_ENCRYPTED"
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertEqual(response.status_code, 200)
        sub_id = response.json()["subscription_id"]
        self.assertIsNotNone(sub_id)

        data["subscription_id"] = sub_id
        for token in invalid_tokens:
            data["token"] = token
            data["push_service"] = "UNIFIED_PUSH_ENCRYPTED"
            response = self.client.put(f'/subscription/?subscription_id={sub_id}', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
            self.assertEqual(response.status_code, 400)

    def test_expire(self):
        prev_count = Subscription.objects.count()
        remove_old_subscription()
        self.assertEqual(Subscription.objects.count(), prev_count)
        subOld = Subscription(
            token="https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1",
            bounding_box=Polygon.from_bbox((8.591, 52.295, 12.063, 52.789)),
            push_service=Subscription.PushServices.UNIFIED_PUSH_ENCRYPTED,
            last_heartbeat=datetime.datetime(2025, 1, 1, 12, 43, 56, 0, datetime.timezone.utc),
            p256dh_key="BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0",
            auth_key="ns9swjbbKTEN12VGW_tJqA",
            vapid_public_key="BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4",
            user_agent="FPAS/1.0.0 (testing)"
        )
        subOld.save()
        subNew = Subscription(
            token="https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1",
            bounding_box=Polygon.from_bbox((8.591, 52.295, 12.063, 52.789)),
            push_service=Subscription.PushServices.UNIFIED_PUSH_ENCRYPTED,
            last_heartbeat=datetime.datetime.now(datetime.timezone.utc),
            p256dh_key="BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0",
            auth_key="ns9swjbbKTEN12VGW_tJqA",
            vapid_public_key="BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4",
            user_agent="FPAS/1.0.0 (testing)"
        )
        subNew.save()
        # Note: the DB fixture for the unit tests has a 10 days expiry
        subJustExpired = Subscription(
            token="https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1",
            bounding_box=Polygon.from_bbox((8.591, 52.295, 12.063, 52.789)),
            push_service=Subscription.PushServices.UNIFIED_PUSH_ENCRYPTED,
            last_heartbeat=datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(days=11),
            p256dh_key="BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0",
            auth_key="ns9swjbbKTEN12VGW_tJqA",
            vapid_public_key="BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4",
            user_agent="FPAS/1.0.0 (testing)"
        )
        subJustExpired.save()
        subAboutToExpire = Subscription(
            token="https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1",
            bounding_box=Polygon.from_bbox((8.591, 52.295, 12.063, 52.789)),
            push_service=Subscription.PushServices.UNIFIED_PUSH_ENCRYPTED,
            last_heartbeat=datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(days=9),
            p256dh_key="BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0",
            auth_key="ns9swjbbKTEN12VGW_tJqA",
            vapid_public_key="BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4",
            user_agent="FPAS/1.0.0 (testing)"
        )
        subAboutToExpire.save()
        self.assertEqual(Subscription.objects.count(), prev_count + 4)
        remove_old_subscription()
        self.assertEqual(Subscription.objects.count(), prev_count + 2)
        self.assertIsNotNone(Subscription.objects.get(id=subNew.id))
        self.assertIsNotNone(Subscription.objects.get(id=subAboutToExpire.id))

    def test_vapid_pub_key(self):
        data = {
            'min_lat': 52.295,
            'max_lat': 52.789,
            'min_lon': 8.591,
            'max_lon': 12.063,
            'p256dh_key': 'BInn4ytZr6wQ960L3sQ6tfmrQzNQoEhj_I-0i2DRcL-_u0aU2vSgLuhLKyzGnFkmKDhfnZ7pwcsOEsqy-fDbzh0',
            'auth_key': 'ns9swjbbKTEN12VGW_tJqA',
            'push_service': 'UNIFIED_PUSH_ENCRYPTED',
            'token': 'https://unifiedpush.kde.org/upezVkNWZjNTM5?up=1',
        }

        # current key is being used as default when not set, for backward compatibility
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertEqual(response.status_code, 200)
        sub_id = response.json()['subscription_id']
        self.assertIsNotNone(sub_id)
        sub = Subscription.objects.get(id=sub_id)
        self.assertEqual(sub.vapid_public_key, 'BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4')

        # update retains existing vapid key
        response = self.client.put(f'/subscription/?subscription_id={sub_id}', json.dumps(data),
                                   content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertEqual(response.status_code, 200)
        sub = Subscription.objects.get(id=sub_id)
        self.assertEqual(sub.vapid_public_key, 'BHJnBOSvBJ9Vl0fF44dUFxmr3l-mNSjuAGvIsFKBSWUsBu2-v2dov1UcGgE2Ry_yjJsz38F3a0A-QrAjCr3OCA4')

        # an explicitly specified but unknown vapid key is rejected
        data['vapid_public_key'] = "SomeUnknownKey"
        response = self.client.post('/subscription/', json.dumps(data), content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertContains(response, "unknown VAPID key", status_code=400)

        # same for updates
        response = self.client.put(f'/subscription/?subscription_id={sub_id}', json.dumps(data),
                                   content_type="application/json", headers={"user_agent": "FPAS/1.0.0 (testing)"})
        self.assertContains(response, "unknown VAPID key", status_code=401)