# openAPI-docu.yaml
# SPDX-FileCopyrightText: Nucleus <nucleus-ffm@posteo.de>
# SPDX-License-Identifier: AGPL-3.0-or-later
openapi: 3.1.0
info:
  title: "FOSS Public Alert Server API"
  version: '1.0.0'
  description: 'This is the API for the FOSS Public Alert Server. The server is in an early testing phase, and the API may change in the near future.'
  termsOfService: 'https://invent.kde.org/webapps/foss-public-alert-server/-/wikis/Terms-of-Service'
  contact:
    name: FOSS Public Alert Server
    url: https://invent.kde.org/webapps/foss-public-alert-server/
    email: nucleus-ffm@posteo.de, vkrause@kde.org

servers:
  - url: https://alerts.kde.org

tags:
  - name: Subscription
  - name: Alerts
  - name: Alert Sources
  - name: Config

paths:
  /subscription/:
    post:
      summary: Subscribe for a new area
      description: Subscribe a client to a UnifiedPush endpoint for an area defined by a bounding box with coordinates.
        As part of subscribing a push notification is sent to the client with a JSON payload with the type set to 'subscribe'
        and `confirmation_id` containing a unique identifier also included in the REST response.
        When new alerts appear in the subscribed area push notifications with a JSON payload with type 'added' are sent which
        contain the identifier of the alert in the 'alert_id' field. Similarly, when an alerts changes in the subscribed area
        a notification with type 'update' is sent.
        A subscriptions needs to be regularly updated to not expire. When a subscription expries a push notification is sent
        with a JSON payload with type 'unsubscribe'.
      tags:
        - Subscription
      requestBody:
        description: Register a new client
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new_subscription'
        required: true
      responses:
        '200':
          description: Client successfully subscribed.
            This implies a successful submission of a push notification to the endpoint given in the request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "successfully subscribed"
                  subscription_id:
                    type: string
                    description: The ID of the subscription to allow the client to unsubscribe. This ID should be stored by the client.
                    example: "exampleSubscriptionID"
                  confirmation_id:
                    type: string
                    description: A unique identifier also present in the push notification sent as a response of this.
        '400':
          description: "Missing required parameters, unsupported or dysfunctional push endpoint, invalid geographic bounding box."
        '405':
          description: "Wrong http method"
    delete:
      summary: Unsubscribe from an area
      description: Unsubscribe a client from an area.
      tags:
        - Subscription
      parameters:
        - in: query
          name: subscription_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Client successfully removed"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "successfully unsubscribed"
        '400':
          description: "Invalid input"
        '405':
          description: "Wrong http method"
        '422':
          description: "Validation exception"

    put:
      summary: Update a subscription
      description: Update a subscription and renew optionally the used tokens
      tags:
        - Subscription
      parameters:
        - in: query
          name: subscription_id
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: false
          schema:
            type: string
      responses:
        '200':
          description: "Subscription successfully updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "successfully updated"
        '400':
          description: "Invalid input"
        '401':
          description: Unknown or invalid VAPID public key, either provided in this request or as part of the existing subscription.
            This can happen as part of VAPID key rotations, clients are expected to retrieve the current VAPID key in this case and
            update the subscription accordingly.
        '404':
          description: "Subscription has expired. You must register again!"
        '405':
          description: "Wrong http method"
        '422':
          description: "Validation exception"

    get:
      summary: fetch push configuration
      description: fetch config for push notifications e.g. the vapid key for webpush
      tags:
        - Subscription
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
          example: "webpush"
      responses:
        '200':
          description: "successfully fetch config"
          content:
            application/json:
              schema:
                type: object
                properties:
                  vapid-key:
                    type: string
                    example: "<vapid key>"
        '400':
          description: "Invalid input"
        '405':
          description: "Wrong http method"
        '422':
          description: "Validation exception"

  /alert/{UUID}:
    get:
      summary: Get one alert
      description: Retrieve a specific alert using the provided alert ID. The alert follows the Common Alerting Protocol (CAP) Version 1.2.
      tags:
        - Alerts
      parameters:
        - in: path
          name: UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "CAP alert data"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/alert_response'
        '400':
          description: "Invalid input"
        '405':
          description: "Wrong http method"
        '422':
          description: "Validation exception"

  /alert/all:
    get:
      summary: Get all alerts for a subscription ID
      description: Retrieve all alerts for a given subscription ID.
      tags:
        - Alerts
      parameters:
        - in: query
          name: subscription_id
          required: true
          schema:
            type: string
          description: "The subscription id"
      responses:
        '200':
          description: "All CAP alerts for the subscription"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/all_alert_response'
        '400':
          description: "No valid subscription"
        '405':
          description: "wrong http method"
        '422':
          description: "Validation exception"

  /alert/area:
    get:
      summary: Get all alerts for an area
      description: Retrieve all alerts for a specified area defined by four coordinates.
      tags:
        - Alerts
      parameters:
        - in: query
          name: min_lat
          required: true
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
          description: "The minimum latitude"
          example: -48.9
        - in: query
          name: max_lat
          required: true
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
          description: "The maximum latitude. Must be larger than `min_lat`."
          example: -48.7
        - in: query
          name: min_lon
          required: true
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
          description: "The minimum longitude"
          example: -123.5
        - in: query
          name: max_lon
          required: true
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
          description: "The maximum longitude. Must be larger than `min_lon`."
          example: -123.3
      responses:
        '200':
          description: "All CAP alerts for the area"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/all_alert_response'
        '400':
          description: "No valid subscription"
        '405':
          description: "wrong http method"
        '422':
          description: "Validation exception"

  /source/area_status:
    get:
      summary: Area status check
      description: Check which feeds the server has for a list of country codes.
      tags:
        - Alert Sources
      parameters:
        - in: query
          name: country_codes
          description:  "A list of country codes in ISO 3166-1 format"
          required: true
          schema:
            type: array
            items:
              type: string
            example: [ "de", "fr" ]
      responses:
        '200':
          description: "List of all feeds for the country codes"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/area_status_response'
        '400':
          description: "Invalid input"
        '405':
          description: "wrong http method"
        '422':
          description: "Validation exception"

  /config/server_status:
    get:
      summary: server status
      description: get information about the operator, urls to privacy police and terms of service, congestion state and supported push services
      tags:
        - Config
      responses:
        '200':
          description: "get server status information"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server_status_response'
        '400':
          description: "Invalid request"
        '405':
          description: "wrong http method"

components:
  schemas:
    new_subscription:
      description: New subscription
      properties:
        token:
          type: string
          description: "The token for sending push notifications for alerts. Depending on the used service e.g. 
          The URL of the UnifiedPush server or the token for APN or Firebase"
          example: ["https://example.de", "bearer<token data in BASE64>"]
        push_service:
          type: string
          description: "Specifies the wanted push notification service"
          example: ["UNIFIED_PUSH", "UNIFIED_PUSH_ENCRYPTED", "APN", "FIREBASE"]
        min_lat:
          type: number
          format: float
          minimum: -90
          maximum: 90
          description: "The minimum latitude"
          example: -48.9
        max_lat:
          type: number
          format: float
          minimum: -90
          maximum: 90
          description: "The maximum latitude. Must be larger than `min_lat`."
          example: -48.7
        min_lon:
          type: number
          format: float
          minimum: -180
          maximum: 180
          description: "The minimum longitude"
          example: -123.5
        max_lon:
          type: number
          format: float
          minimum: -180
          maximum: 180
          description: "The maximum longitude. Must be larger than `min_lon`, ie. the subscribed area cannot cross the antimeridian."
          example: -123.3
        p256dh_key:
          type: string
          description: EDCH user agent public key according to RFC 8291, in Base64 URL encoding. (Only for UNIFIED_PUSH_ENCRYPTED).
        auth_key:
          type: string
          description: 16 byte authentication secret according to RFC 8291, in Base64 URL encoding. (Only for UNIFIED_PUSH_ENCRYPTED).
        vapid_public_key:
          type: string
          description: VAPID public key used by the client, in Base64 URL encoding. (Only for UNIFIED_PUSH_ENCRYPTED).


    alert_request:
      description: "Request a specific alert by ID"
      properties:
        alert_id:
          type: string
          description: "The UUID of the alert"
          example: "ExampleAlertUUID"

    alert_response:
      type: object
      properties:
        identifier:
          type: string
          example: "urn:oid:2.49.0.1.4.0.2011.2.2.8.37.9"
        sender:
          type: string
          example: "geomag@usgs.gov"
        sent:
          type: string
          format: date-time
          example: "2011-02-02T08:37:09-00:00"
        status:
          type: string
          example: "Test"
        msgType:
          type: string
          example: "Alert"
        scope:
          type: string
          example: "Public"
        info:
          type: object
          properties:
            language:
              type: string
              example: "en"
            category:
              type: string
              example: "Met"
            event:
              type: string
              example: "Geomagnetic Storm"
            urgency:
              type: string
              example: "Expected"
            severity:
              type: string
              example: "Moderate"
            certainty:
              type: string
              example: "Likely"
            senderName:
              type: string
              example: "USGS Geomagnetism Program"
            headline:
              type: string
              example: "Geomagnetic Storm Alert"
            description:
              type: string
              example: "The U.S. Geological Survey (USGS) is bracing for the likely arrival of a major geomagnetic storm..."
            instruction:
              type: string
              example: "While geomagnetic storms give rise to the beautiful Northern lights..."
            web:
              type: string
              example: "http://www.usgs.gov/public/press/public_affairs/press_releases/pr1241m.html"
            contact:
              type: string
              example: "Jill Caldwell 1-303-273-8486"
            resource:
              type: object
              properties:
                resourceDesc:
                  type: string
                  example: "Image file"
                mimeType:
                  type: string
                  example: "image/jpeg"
                uri:
                  type: string
                  example: "http://geomag.usgs.gov/realtime.html"
            area:
              type: object
              properties:
                areaDesc:
                  type: string
                  example: "Plots of the data from USGS observatories can be seen on-line at..."
                polygon:
                  type: string
                  example: "45,-179.99 45,179.99 89.99,179.99 89.99,-179.99 45,-179.99"


    area_status_response:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Deutschland: Landeshochwasserportal"
              source_is_official:
                type: boolean
                example: true
              cap_alert_feed_status:
                type: string
                example: "operating"
              authorityCountry:
                type: string
                example: "de"
              register_url:
                type: string
                example: "https://alertingauthority.wmo.int/authorities.php?recId=53"
              latest_published_alert_datetime:
                type: string
                example: "2024-12-10T05:00:44Z"

    all_alert_response:
      description: "All alerts for the given subscription"
      type: array
      items:
        $ref: '#/components/schemas/alert_response'

    server_status_response:
      description: "information about the sever settings"
      properties:
        server_version:
          type: string
          description: "The installed server version and build revision"
          example: "1.0.0 (abcde)"
        server_operator:
          type: string
          description: "Information about how operates the server"
          example: "Example Operator"
        privacy_notice:
          type: string
          description: "URL to the privacy police"
          example: "https://example.com/privacy"
        terms_of_service:
          type: string
          description: "url to the terms of service"
          example:  "https://example.com/tos"
        congestion_state:
          type: number
          format: int
          description: "overload indicator for the server 1: everything fine - 10: no more subscriptions possible"
          example: 1
        supported_push_services:
          type: object
          additionalProperties:
            type: boolean
          properties:
            UNIFIED_PUSH:
              type: boolean
            UNIFIED_PUSH_ENCRYPTED:
              type: boolean
            APN:
              type: boolean
            FIREBASE:
              type: boolean
          example:
              UNIFIED_PUSH: true
              UNIFIED_PUSH_ENCRYPTED: false
              APN: false
              FIREBASE: false